10_quantization/requirements.txt
huggingface_hub
torch
transformers
llama-cpp-python
